+++ CHANGELOG.md
@@ -14,10 +14,8 @@ nav_order: 1
 - Change Flink version to 1.15
 - Fix empty `user_config` and empty `ip_filters`
 - Fix `tools/` consistency
-
-## [3.6.1] - 2022-09-05
-
 - Add support `is_account_owner` [field](https://api.aiven.io/doc/#operation/AccountList)
+- send force user_config on service integration update, add `datadog_user_config`, resolves #888
 
 ## [3.6.0] - 2022-08-31